Make animation steps not lose time due to delays elsewhere.
last_timer is now the time that the previous animation step was
scheduled to occur, rather than when it actually did occur. This means
that overall each animation step will proceed according to scheduled times.

Side effect: Currently, when you switch to another animation (or disable
animations) and then return to the animation, it runs many steps as fast
as it can in order to "catch up" to where it should be.